LayoutTests/css3/scroll-snap/scroll-snap-desination-lock-up.html

<!DOCTYPE html>
<html>
    <head>
        <style>
            .horizontalGallery {
                width: 400px;
                height: 50px;
                overflow-y: hidden;
                overflow-x: auto;
            }
            .horizontalGalleryDrawer {
                width: 3000px;
                height: 50px;
            }
            .colorBox {
                height: 50px;
                width: 400px;
                float: left;
            }

            .lockUpSnapCoordinate {
                -webkit-scroll-snap-type: mandatory;
                -webkit-scroll-snap-points-x: repeat(100%);
                -webkit-scroll-snap-points-y: repeat(100%);
                -webkit-scroll-snap-destination: 100% 100%;
                -webkit-scroll-snap-coordinate: initial;
            }

            #item0 { background-color: red; }
            #item1 { background-color: green; }
            #item2 { background-color: blue; }
            #item3 { background-color: aqua; }
            #item4 { background-color: yellow; }
            #item5 { background-color: fuchsia; }
        </style>
        <script src="../../resources/js-test-pre.js"></script>
    </head>
    <body>
        <div style="width: 400px">
            <div id="lockUpSnapCoordinate" class="horizontalGallery lockUpSnapCoordinate">
                <div class="horizontalGalleryDrawer">
                    <div id="item0" class="colorBox"></div>
                    <div id="item1" class="colorBox"></div>
                    <div id="item2" class="colorBox"></div>
                    <div id="item3" class="colorBox"></div>
                    <div id="item4" class="colorBox"></div>
                    <div id="item5" class="colorBox"></div>
                </div>
            </div>
        </div>
        This test passes if it doesn't time out.
        <div id="console"></div>
        <script>
        description("Tests infinite loop calculating scroll snap point offsets.");

        var lockUpSnapCoordinate = document.getElementById('lockUpSnapCoordinate');
        shouldBeEmptyString("lockUpSnapCoordinate.style['-webkit-scroll-snap-type']");
        shouldBeEqualToString("window.getComputedStyle(lockUpSnapCoordinate).getPropertyValue('-webkit-scroll-snap-type')", "mandatory");
        shouldBeEmptyString("lockUpSnapCoordinate.style['-webkit-scroll-snap-points-x']");
        shouldBeEqualToString("window.getComputedStyle(lockUpSnapCoordinate).getPropertyValue('-webkit-scroll-snap-points-x')", "repeat(100%)");
        shouldBeEmptyString("lockUpSnapCoordinate.style['-webkit-scroll-snap-points-y']");
        shouldBeEqualToString("window.getComputedStyle(lockUpSnapCoordinate).getPropertyValue('-webkit-scroll-snap-points-y')", "repeat(100%)");
        shouldBeEmptyString("lockUpSnapCoordinate.style['-webkit-scroll-snap-destination']");
        shouldBeEqualToString("window.getComputedStyle(lockUpSnapCoordinate).getPropertyValue('-webkit-scroll-snap-destination')", "100% 100%");
        shouldBeEmptyString("lockUpSnapCoordinate.style['-webkit-scroll-snap-coordinate']");
        shouldBeEqualToString("window.getComputedStyle(lockUpSnapCoordinate).getPropertyValue('-webkit-scroll-snap-coordinate')", "none");
        </script>
        <script src="../../resources/js-test-post.js"></script>
    </body>
</html>
