LayoutTests/css3/scroll-snap/scroll-snap-desination-lock-up-expected.txt

This test passes if it doesn't time out.
Tests infinite loop calculating scroll snap point offsets.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS lockUpSnapCoordinate.style['-webkit-scroll-snap-type'] is ""
PASS window.getComputedStyle(lockUpSnapCoordinate).getPropertyValue('-webkit-scroll-snap-type') is "mandatory"
PASS lockUpSnapCoordinate.style['-webkit-scroll-snap-points-x'] is ""
PASS window.getComputedStyle(lockUpSnapCoordinate).getPropertyValue('-webkit-scroll-snap-points-x') is "repeat(100%)"
PASS lockUpSnapCoordinate.style['-webkit-scroll-snap-points-y'] is ""
PASS window.getComputedStyle(lockUpSnapCoordinate).getPropertyValue('-webkit-scroll-snap-points-y') is "repeat(100%)"
PASS lockUpSnapCoordinate.style['-webkit-scroll-snap-destination'] is ""
PASS window.getComputedStyle(lockUpSnapCoordinate).getPropertyValue('-webkit-scroll-snap-destination') is "100% 100%"
PASS lockUpSnapCoordinate.style['-webkit-scroll-snap-coordinate'] is ""
PASS window.getComputedStyle(lockUpSnapCoordinate).getPropertyValue('-webkit-scroll-snap-coordinate') is "none"
PASS successfullyParsed is true

TEST COMPLETE
